refactor(creator): collapse store repetition, split InspectorPanel, adopt EffectListEditor
Item 4 of the apps quality review: the Creator store and InspectorPanel
had accreted mechanical repetition without being god objects.

store/index.ts (13x + 4x repeated tails -> two helpers):
- commitDoc() collapses the derive-RF -> revalidate -> annotate-errors ->
  set tail copy-pasted across syncFromRF, addNode, deleteNode,
  updateNodeProps, commitDungeons, commitBoards, setActiveBoard,
  addDungeonSubflow, createGroup, setEntry, and others. Accepts optional
  pre-derived nodes/edges (syncFromRF needs its own RF state, which
  carries selection/drag runtime fields a fresh derive would drop) and an
  optional mapNodes hook (setEntry's isEntry stamp).
- commitDocOnly() collapses the revalidate-only tail for scenario-wide,
  non-node-scoped edits (updateScenarioDescription, updateSetupSelections,
  updateMainGoal).
- commitLibrary()/setOrDeleteKey() collapse the clone-library ->
  set-or-delete-if-empty -> revalidate -> set tail across
  updateLibraryCards/updateLibraryDecks/updateBattleDefs/
  updateResourceImages/updateFoeBattleDefId.

Node-error attribution (the store's error->node mapping and Inspector's
node->error filter disagreed: a node could show an error badge on canvas
the Inspector didn't list). Both now go through one shared primitive,
utils/nodeErrors.ts (errorMentionsNode/nodeIdsInError) -- a correctness
fix, not just dedup: the store's old regex grabbed the first quoted token
in an error string with no check it was an actual node id.

InspectorPanel.tsx (1208 -> ~370 lines): the nine components already
living in one file are now separate leaf files.
- components/CollapsibleSection.tsx: the generic UI primitive.
- editors/inspector/*.tsx: DungeonSubflowEditor, TowerOpEditor,
  GroupEditor, ScenarioSetupEditor, BoardSetupEditor, SelectHeroEditor.
- editors/inspector/registry.ts: a NODE_EDITORS
  Record<NodeKind, ComponentType> lookup replaces the hand-rolled
  sn.kind === 'x' && <XEditor/> chain for the five node-kind editors
  sharing one render slot (util.group keeps its separate slot/order
  unchanged).
- utils/objectHelpers.ts: asObj, defined identically three times across
  the file, collapsed to one (asStr stays local -- it was only ever
  defined once).
- EffectApplyEditor now renders the shared, controlled EffectListEditor
  (deck builder / dungeon editor's editor) instead of a raw JSON
  textarea -- the file's clearest accretion tell per the review: the
  good abstraction existed and every other surface had already adopted
  it. Verified live in the browser (structured op/resource/amount
  fields, correct store round-trip, no console errors) plus a rewritten
  InspectorPanel.test.tsx pinning that a selection change never leaks
  one node's effects onto another (the fix this replaces a workaround
  for).

Verified with pnpm run ci (validate:nodes, lint, format, build,
typecheck, test -- 76/76 creator tests) and a live creator session
(effect.apply structured editing, tower.op, and an unregistered kind all
render correctly with clean console). apps/creator is private; no
changeset needed.
