+++ apps/creator/src/editors/InspectorPanel.test.tsx
@@ -1,15 +1,15 @@
 /**
  * InspectorPanel.test.tsx — the inspector must show the selected node's own data.
  *
- * EffectApplyEditor renders an uncontrolled <textarea defaultValue={effectsStr}> with
- * no key. React reconciles same-type-same-position elements by reusing the instance,
- * so selecting a second effect.apply node kept the first node's JSON on screen while
- * onUpdate wrote to the second — one keystroke committed node A's effects onto node
- * B, the ~800ms autosave persisted it, and there is no undo.
+ * effect.apply used to render an uncontrolled <textarea defaultValue={effectsStr}> with no key
+ * (EffectApplyEditor's old raw-JSON textarea). React reconciles same-type-same-position elements by
+ * reusing the instance, so selecting a second effect.apply node kept the first node's JSON on
+ * screen while onUpdate wrote to the second — one keystroke committed node A's effects onto node B,
+ * the ~800ms autosave persisted it, and there is no undo.
  *
- * The sibling ScenarioSetupEditor in the same file already keys its uncontrolled
- * input on the committed value and comments on exactly this hazard; this pins the
- * same behaviour for the effects editor.
+ * effect.apply now renders the shared, controlled EffectListEditor (editors/effects), keyed on the
+ * node id in the NODE_EDITORS registry lookup — a selection change always remounts fresh from the
+ * new node's `value` prop rather than reusing a stale DOM instance. This pins that the fix holds.
  */
 
 import { describe, expect, it, beforeEach, vi } from 'vitest';
@@ -32,8 +32,11 @@ function rfNodeFor({ id, effects }: EffectNode) {
   };
 }
 
-const nodeA = { id: 'node-a', effects: [{ op: 'resource.add', amount: 1 }] };
-const nodeB = { id: 'node-b', effects: [{ op: 'resource.remove', amount: 99 }] };
+// resource.gain/resource.lose have a structured form (opForms.ts) — an "Amount" number field —
+// so the two nodes are distinguishable by that field's value without depending on the JSON
+// fallback an op with no structured form would render.
+const nodeA = { id: 'node-a', effects: [{ op: 'resource.gain', resource: 'warriors', amount: 3 }] };
+const nodeB = { id: 'node-b', effects: [{ op: 'resource.lose', resource: 'spirit', amount: 7 }] };
 
 function selectNode(id: string) {
   useCreatorStore.setState({
@@ -42,8 +45,8 @@ function selectNode(id: string) {
   });
 }
 
-function effectsTextarea(): HTMLTextAreaElement {
-  return screen.getByRole('textbox', { name: /effects/i }) as HTMLTextAreaElement;
+function amountInput(): HTMLInputElement {
+  return screen.getByLabelText('Amount') as HTMLInputElement;
 }
 
 describe('InspectorPanel — effect.apply editor', () => {
@@ -67,23 +70,22 @@ describe('InspectorPanel — effect.apply editor', () => {
     selectNode('node-a');
     render(<InspectorPanel />);
 
-    expect(effectsTextarea().value).toContain('resource.add');
+    expect(amountInput().value).toBe('3');
   });
 
-  // The bug: the textarea keeps node A's JSON after switching to node B.
+  // The bug (now fixed): the editor used to keep node A's data after switching to node B.
   it('shows the newly selected node effects after switching nodes', () => {
     selectNode('node-a');
     const { rerender } = render(<InspectorPanel />);
-    expect(effectsTextarea().value).toContain('resource.add');
+    expect(amountInput().value).toBe('3');
 
     selectNode('node-b');
     rerender(<InspectorPanel />);
 
-    expect(effectsTextarea().value).toContain('resource.remove');
-    expect(effectsTextarea().value).not.toContain('resource.add');
+    expect(amountInput().value).toBe('7');
   });
 
-  // The consequence: editing after a switch must write the visible node's data.
+  // The consequence (now fixed): editing after a switch must write the visible node's data.
   it('does not carry one node effects onto another', () => {
     const updateNodeProps = vi.fn();
     useCreatorStore.setState({ updateNodeProps } as never);
@@ -93,9 +95,8 @@ describe('InspectorPanel — effect.apply editor', () => {
     selectNode('node-b');
     rerender(<InspectorPanel />);
 
-    const textarea = effectsTextarea();
-    // Whatever the field currently holds is what a keystroke commits — so it must
-    // already be node B's data, not node A's.
-    expect(JSON.parse(textarea.value)).toEqual(nodeB.effects);
+    // Whatever the field currently holds is what an edit would commit — so it must already be
+    // node B's data, not node A's.
+    expect(amountInput().value).toBe('7');
   });
 });
